+++ tsr-rsc/src/integrations/flight/ssr.ts
@@ -1,3 +1,4 @@
+import { tinyassert } from "@hiogawa/utils";
 import { $__global } from "../global";
 import { streamToString } from "../utils";
 import type { FlightData } from "./client";
@@ -9,6 +10,19 @@ export async function handleFlight(
 	const reactServer = await importReactServer();
 	const stream = await reactServer.handler(reference, args);
 	return { __flight: await streamToString(stream) };
+}
+
+export async function handleFlightRequest(request: Request) {
+	const url = new URL(request.url);
+	const reference = url.searchParams.get("reference");
+	tinyassert(reference);
+	const args = await request.json();
+	const data = await handleFlight(reference, args);
+	return new Response(JSON.stringify(data), {
+		headers: {
+			"content-type": "application/json",
+		},
+	});
 }
 
 export async function importReactServer(): Promise<typeof import("./server")> {
